small fix in histograms
small fix in histograms to avoid null

+++ src/offlineProcessing/OfflineTask.java
@@ -383,6 +383,9 @@ public abstract class OfflineTask<T extends PamDataUnit> {
 		}
 
 		for (PamDataBlock aBlock:affectedDataBlocks) {
+			if (aBlock == parentDataBlock) {
+				continue;
+			}
 			deleteOldData(aBlock, taskGroupParams);
 		}
 	}
